+++ pkg/parse/radiobridge/door_test.go
@@ -22,39 +22,39 @@ func TestDoor(t *testing.T) {
 	}{
 		// Reset.
 		{"100001120102181c", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 0},
-			{Attr: "hw_ver", Value: 18},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(0)},
+			{Attr: "hw_ver", Value: int32(18)},
 			{Attr: "ver", Value: "1.2"},
 		}, ""},
 		// Supervisory.
 		{"1401080131", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 4},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(4)},
 			{Attr: "tamper", Value: false},
 			{Attr: "battery", Value: 3.1},
 		}, ""},
 		// Tamper.
 		{"1c0200", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 12},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(12)},
 			{Attr: "tamper", Value: true},
 		}, ""},
 		// Link Quality.
 		{"1dfb010000", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 13},
-			{Attr: "sub_band", Value: 1},
-			{Attr: "dev_rssi", Value: 0},
-			{Attr: "dev_snr", Value: 0},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(13)},
+			{Attr: "sub_band", Value: int32(1)},
+			{Attr: "dev_rssi", Value: int32(0)},
+			{Attr: "dev_snr", Value: int32(0)},
 		}, ""},
 		// Door.
 		{"190301", []*parse.Point{
-			{Attr: "count", Value: 9},
+			{Attr: "count", Value: int32(9)},
 			{Attr: "open", Value: true},
 		}, ""},
 		{"1a0300", []*parse.Point{
-			{Attr: "count", Value: 10},
+			{Attr: "count", Value: int32(10)},
 			{Attr: "open", Value: false},
 		}, ""},
 		// Door bad length.
@@ -63,7 +63,7 @@ func TestDoor(t *testing.T) {
 		{"190401", nil, "door format bad identifier: 190401"},
 		// Door unused trailing bytes.
 		{"190301ff", []*parse.Point{
-			{Attr: "count", Value: 9},
+			{Attr: "count", Value: int32(9)},
 			{Attr: "open", Value: true},
 		}, "door format unused trailing bytes: 190301ff"},
 	}
