+++ pkg/parse/radiobridge/supervisory_test.go
@@ -22,14 +22,14 @@ func TestSupervisory(t *testing.T) {
 	}{
 		// Supervisory.
 		{"1401080131", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 4},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(4)},
 			{Attr: "tamper", Value: false},
 			{Attr: "battery", Value: 3.1},
 		}, ""},
 		{"1401170127", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 4},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(4)},
 			{Attr: "error", Value: "radio_comm"},
 			{Attr: "error", Value: "battery_low"},
 			{Attr: "error", Value: "last_downlink"},
@@ -39,18 +39,18 @@ func TestSupervisory(t *testing.T) {
 		}, ""},
 		// Supervisory event count.
 		{"1701080130ffffffff1234", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 7},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(7)},
 			{Attr: "tamper", Value: false},
 			{Attr: "battery", Value: float64(3)},
-			{Attr: "total_count", Value: 4660},
+			{Attr: "total_count", Value: int32(4660)},
 		}, ""},
 		{"1401080132000000000002", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 4},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(4)},
 			{Attr: "tamper", Value: false},
 			{Attr: "battery", Value: 3.2},
-			{Attr: "total_count", Value: 2},
+			{Attr: "total_count", Value: int32(2)},
 		}, ""},
 		// Supervisory bad length.
 		{"", nil, "supervisory format bad length: "},
@@ -58,16 +58,16 @@ func TestSupervisory(t *testing.T) {
 		{"1402080131", nil, "supervisory format bad identifier: 1402080131"},
 		// Supervisory bad error bitmap.
 		{"1401400131", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 4},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(4)},
 		}, "supervisory format bad error bitmap: 1401400131"},
 		// Supervisory event count unused trailing bytes.
 		{"1701080130ffffffff1234ff", []*parse.Point{
-			{Attr: "proto", Value: 1},
-			{Attr: "count", Value: 7},
+			{Attr: "proto", Value: int32(1)},
+			{Attr: "count", Value: int32(7)},
 			{Attr: "tamper", Value: false},
 			{Attr: "battery", Value: float64(3)},
-			{Attr: "total_count", Value: 4660},
+			{Attr: "total_count", Value: int32(4660)},
 		}, "supervisory format unused trailing bytes: " +
 			"1701080130ffffffff1234ff"},
 	}
